Update ioreg with new test files and update README

+++ cm/hal_ioreg.c
@@ -3,10 +3,10 @@
 
 
 
-u8 hal_IOReg_Read(u8 port) {
-    return bsl_IOReg_Read(port);
+u32 hal_IOReg_Read(u32 port) {
+    return bsl_IOReg_Read((u8)port);
 }
 
-void hal_IOReg_Write(u8 port, u8 value) {
-    bsl_IOReg_Write(port, value);
+void hal_IOReg_Write(u32 port, u32 value) {
+    bsl_IOReg_Write((u8)port, (u8)value);
 }
